+++ src/transcribers/liveProtocol.ts
@@ -50,7 +50,11 @@ export function setupMessage(
 		inputAudioTranscription,
 		realtimeInputConfig: {
 			automaticActivityDetection: {
-				disabled: true,
+				disabled: false,
+				startOfSpeechSensitivity: "START_SENSITIVITY_HIGH",
+				endOfSpeechSensitivity: "END_SENSITIVITY_LOW",
+				silenceDurationMs: 2000,
+				prefixPaddingMs: 300,
 			},
 		},
 	};